Move PrecompiledRegexes in shopping basket component to helper namespace and update usage in ShoppingBasketsService
